Jenkin's file fix/mbed-os.lib update
Ublox C030 and IAR toolchain is taken out temporarily.
Will be added later on.

mbed-os.lib is updated.

Help from the mbed_app.json is removed as we will be parsing that string
in Jenkins file to replace the platform runtime. Just for the ease of operation,
we have removed the help. Documentation will be added instead.